backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.3",
    "build==1.4.0",
    "certifi==2026.2.25",
    "cffi==2.0.0 ; platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.5",
    "chromadb==1.5.5",
    "click==8.3.1",
    "colorama==0.4.6 ; os_name == 'nt' or sys_platform == 'win32'",
    "cryptography==46.0.5",
    "cuda-bindings==12.9.4 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "cuda-pathfinder==1.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "durationpy==0.10",
    "fastapi>=0.135.1",
    "filelock==3.25.1",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "google-auth==2.49.0",
    "google-genai==1.66.0",
    "googleapis-common-protos==1.73.0",
    "greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "hf-xet==1.3.2 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.6.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==35.0.0",
    "langchain==1.2.10",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.2",
    "langchain-community==0.4.1",
    "langchain-core==1.2.18",
    "langchain-google-genai==4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface==1.2.1",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.10",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.10",
    "langsmith==0.7.16",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mmh3==5.2.1",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "networkx==3.6.1",
    "numpy==2.4.3",
    "nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nccl-cu12==2.27.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvshmem-cu12==3.4.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "oauthlib==3.3.1",
    "onnxruntime==1.24.3",
    "opentelemetry-api==1.40.0",
    "opentelemetry-exporter-otlp-proto-common==1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.40.0",
    "opentelemetry-proto==1.40.0",
    "opentelemetry-sdk==1.40.0",
    "opentelemetry-semantic-conventions==0.61b0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.3",
    "pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pygments==2.19.2",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==5.2.3",
    "setuptools==82.0.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "sqlalchemy==2.0.48",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.2",
    "torch==2.10.0",
    "tqdm==4.67.3",
    "transformers==5.3.0",
    "triton==3.6.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.41.0",
    "uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
